add tests

+++ src/main/java/com/frobbery/gamification/service/level/LevelService.java
@@ -5,7 +5,7 @@ import com.frobbery.gamification.util.dto.LevelDto;
 
 public interface LevelService {
 
-    int getAllAvailableNum();
+    long getAllAvailableNum();
 
     int getLastOpenNum(String userEmail);
